futility: updater: allow --fast to skip scanning archive in do_manifest
The `--manifest` (do_manifest) command needs to scan most firmware
images in the archive to build up the right information. That can be ~2s
for a 200MB archive even though we just need the version string. 2s is
usually fast enough for developers when they want to check the archive
contents, but that is too slow if the boot time scripts (for example,
CSME updater) need to collect the information.

As a result, we want to allow overriding how the updater gets the
manifest. For most systems that the firmware archive is created by the
buildbot or the signerbot, the scripts can use "--manifest --fast" to
retrieve the cached JSON manifest file.

BUG=b:238908603
TEST=make; run test
BRANCH=None

Change-Id: I5d7dead4d0a43129fd31dd035aac63deaee42a08
Signed-off-by: Hung-Te Lin <hungte@chromium.org>