firmware: 2sha: Add tryhw digest API to manage HW crypto support
This patch adds a new family of functions vb2_digest_init_tryhw(),
vb2_hash_calculate_tryhw() and vb2_hash_verify_tryhw() equivalent to the
similarly named existing hash functions. The difference is that these
functions will first try to call the vb2ex_hwcrypto routines to use
cryptographic acceleration for the calculation before falling back to
the software implementation. Callers can also use an extra argument to
indicate whether using HW crypto is currently disallowed by policy (this
should usually be the value returned by vb2api_hwcrypto_allowed()). When
disallowed, the functions will force using the software implementation
(this allows deduplicating the usual fallback logic and logging around
this case).

Also export vb2_hwcrypto_allowed() as a vb2api function to make it
available to external callers.

Finally, make an API requirements change to the vb2ex_hwcrypto API to
allow the data_size value to vb2ex_hwcrypto_digest_init() to be 0. This
is necessary because callers cannot always know the total data size in
advance, and not all HW crypto implementations actually need to know the
size, so we shouldn't preclude using HW crypto for those that don't in
cases where it cannot be known just because others do. The
implementations that do need to know the size will now need to check
whether the value is 0 and return HWCRYPTO_UNSUPPORTED in that case.

BRANCH=none
BUG=b:240624460
TEST=runtests

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I34c3f54e31742619d422d1cd871bdb77ad0439b7