futility: Don't verify checksums in validate_rec_mrc
CB:67670 replaced the header and data checksums in the MRC header with
hashes, so `futility validate_rec_mrc` will fail when validating the
checksums.

Remove the checksum verification from validate_rec_mrc. This command is
only used by the MrcCacheTest factory test. The goal of this test is to
check that the MRC cache is repopulated after clearing it (see
b:35582652 for more context on the original motivation). Checking that
the signature is correct should be enough to verify this.

We already have FAFT tests (firmware_CorruptRecoveryCache,
firmware_RecoveryCacheBootKeys) which check the functionality of the
hash, e.g. check that memory is retrained when the cache is corrupted.

Also update the mrc_metadata struct to match CB:67670.

BUG=b:242667207,b:245277259,b:246302358
TEST=`futility validate_rec_mrc` succeeds on nivviks
BRANCH=None

Change-Id: I66a17124061a9c36852dbfba3575dd7add0beea8
Signed-off-by: Reka Norman <rekanorman@google.com>