futility: Don't verify data checksum in validate_rec_mrc
coreboot has two different ways of verifying the MRC cache data:
1. If MRC_SAVE_HASH_IN_TPM is selected, the data is hashed and verifed
   against a hash stored in the TPM.
2. Otherwise, an IP checksum is calculated and compared against a
   checksum stored in the MRC cache header.

Previously, the checksum was calculated and stored in the header even
when option 1 is used, even though nothing uses it. CB:67042 removed
this unnecessary calculation to save boot time. However, now `futility
validate_rec_mrc` fails when verifying the data checksum.

Remove the data checksum verification from validate_rec_mrc. This
command is only used by the MrcCacheTest factory test. The goal of this
test is to check that the MRC cache is repopulated after clearing it
(see b:35582652 for more context on the original motivation). The other
checks done by validate_rec_mrc (verifying the signature and header
checksum) should be enough to verify this.

We already have FAFT tests (firmware_CorruptRecoveryCache,
firmware_RecoveryCacheBootKeys) which check the functionality of the
checksum or hash, e.g. check that memory is retrained when the cache is
corrupted.

BUG=b:242667207,b:245277259
TEST=`futility validate_rec_mrc` succeeds on nivviks
BRANCH=None

Change-Id: I66a17124061a9c36852dbfba3575dd7add0beea8
Signed-off-by: Reka Norman <rekanorman@google.com>