futility: updater: Add support for per-SKU_ID image-name overrides.
Now that models can use different firmware image names for different
SKUs, special handling is needed in cases where crosid is unavailable,
such as firmware provisioning over servo.

Add a --image_name option to futility update to explicitly specify the
image-name to use, equivalent to specifying the firmware manifest key
returned by crosid.

If instead the model name and SKU ID are available, consult an override
CSV table in models/${model}/image_name_overrides.csv, searching for an
entry with the specified SKU ID. If a matching entry is found, use its
image name instead of the specified model since that model, SKU ID
combination requires a different image name to be used.

If no overrides file is present, or a SKU ID is not specified, use the
model name as the image-name.

BUG=b:229022567, b:243470230
TEST=futility update --servo -a /path/to/firmware --model nereid \
     --sku_id 0x30001
BRANCH=None

Change-Id: If4ac9710e7da5018088502e1161b951529fe1481
Signed-off-by: Sam McNally <sammc@chromium.org>