devkeys: rename arv root key
Change the name of the generated AP RO verification root key file to
match the name generated on the HSM. Generate a new root key and
re-sign the platform key to make sure that git history shows proper
provenance of the platform keyblock.

The keyblock was generated as follows:

$ futility vbutil_keyblock --pack tests/devkeys/arv_platform.keyblock \
         --datapubkey tests/devkeys/arv_platform.vbpubk \
	 --signprivate tests/devkeys/root_key_arv_root.vbprivk

Also modified the top comment to pacify the preupload check.
BRANCH=none
BUG=b:247652363
TEST=used the new keys to sign and verify a Nissa AP firmware image

Change-Id: Ibe73472a65eb78c118ff08a642c1e111914874b5
Signed-off-by: Vadim Bendebury <vbendeb@google.com>