sign_official_build: add AP RO signing
When signing AP RO images, in cases when signer_config.csv manifest
includes the brand code column add a futility invocation to sign the
RO_GSCVD section of the image. If the <path to unpacked>/keyset
directory is found, save the gscvd.<model> blob in that directory.

BRANCH=none
BUG=b:247652363
TEST=built ChromeOS test image for Nissa, then invoked
  $ scripts/image_signing/sign_official_build.sh \
   base \
   ~/trunk/src/build/images/nissa/latest/chromiumos_test_image.bin \
   tests/devkeys \
   /tmp/signed.bin

  and observed 'futility gscvd' invocation in the log.

Cq-Depend: 3954963
Change-Id: I55cec75794560662ed2cfb2dac7f44d972a8571f
Signed-off-by: Vadim Bendebury <vbendeb@google.com>