futility: Remove validate_rec_mrc command
The `futility validate_rec_mrc` command was added to allow the
MrcCacheTest factory test to verify the MRC cahce. However,
duplicating the verification logic between coreboot and futility is not
very maintainable. This caused problems recently when CB:67670 replaced
the checksum in the MRC header with a hash.

MrcCacheTest was changed to instead read eventlog to check that coreboot
successfully updated the cache: CL:3939403. Now `futility
validate_rec_mrc` is no longer used, so remove it.

BRANCH=none
BUG=b:242667207, b:249174725
TEST=emerge-nissa vboot_reference

Change-Id: I7b78cdbfdda37e70aef3675636ed347a4ea1694e
Signed-off-by: Reka Norman <rekanorman@google.com>