futility: add option to save ro_gscvd section in a blob
To support the "futility update" ability to put together firmware
images for different targets, the pre-signed RO_GSCVD sections need to
be included in the firmware tarball.

This patch adds a command line option which will make 'futilty gscvd'
save the signed section in a local file for inclusion in the tarball.

BRANCH=none
BUG=b:245799496
TEST=verified that passing the --gscvd_out command line option results
     in creating a file with the RO_GSCVD section contents, identical
     to the section in the signed binary file.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Id4a75c74ad1e27c11a6005472708730b8051f036