futility: updater: Detect the model via FRID for non-host programmers
When updating with --archive and a non-host programmer (and thus no
reliable crosid to discover the appropriate firmware manifest key), and
no explicit --model parameter is passed, try to detect the model by
matching the FRID of the current firmware with one of the host firmware
images in the archive.

Add a --detect-model-only flag to perform the same matching, but report
the detected model name and exit. This can be used in combination with
the manifest to automatically select an appropriate EC image to pass to
flash_ec.

BUG=b:253966060
TEST=futility update -a firmware.tar.bz2 --servo
BRANCH=None

(cherry picked from commit ff29ee63ed52b698afcaa4c2619d7163322a3785)

Original-Signed-off-by: Sam McNally <sammc@chromium.org>
Change-Id: I25fa0f109d0d8052179b220251d4720438b93bc4
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3965584
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
GitOrigin-RevId: ff29ee63ed52b698afcaa4c2619d7163322a3785
Cr-Build-Id: 8739906983328173473
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739906983328173473
Copybot-Job-Name: android-main-fw-nissa-vboot_reference-copybot-downstream