futility: updater: Add --detect-model(-only) flags.
Add --detect-model to try to detect the model for update --archive
invocations by matching the FRID of the current firmware with one of the
host firmware images in the archive. If a non-default programmer is
selected by command-line, either directly via --programmer or indirectly
via --servo etc., and --archive is provided, enable --detect-model
unless --model is passed. When not using the default host programmer,
the crosid-derived identity is never correct.

Add --detect-model-only to do the same but report the detected model
name and exit. This can be used in combination with the manifest to
automatically select an appropriate EC image to pass to flash_ec.

BUG=b:253966060
TEST=futility update -a firmware.tar.bz2 --detect-model --servo
BRANCH=None

Signed-off-by: Sam McNally <sammc@chromium.org>
Change-Id: I25fa0f109d0d8052179b220251d4720438b93bc4