sign_official_build: do not fail if AP RO signing is not needed
Most of AP firmware images do not include the RO_GSCVD section and are
not supposed to be signed for AP RO verification.

The presence of AP RO verification keys (files prefixed with arv_...)
can be considered an indicator of the need to sign the RO_GSCVD
section.

This patch adds logic to skip signing of AP RO in case the appropriate
signing keys are not present.

BRANCH=none
BUG=b:247645824, chromium:1382709
TEST=ran sign_official_build.sh to re-sign a Nissa test tarball,
      observed successful completion with log messages confirming
      RO_GSCVD signing.

      then removed tests/devkeys/arv_root.vbpubk and ran the script
      again, observed successful completion and log messages
      confirming skipping AP RO verification signing.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Iee5a2adcceb7ecc86f48d7c56755cc10405e5eed