crossystem: Add board_id property
futility is one of a few places in ChromeOS that uses "mosys platform
version".  The goal is to remove this command from mosys.

This commit adds a new property to crossystem, "board_id", which
reads the board revision from SMBIOS/FDT, and replaces the call in
futility with the appropriate VbGetSystemPropertyInt.

BUG=b:187790074
BRANCH=none
TEST="crossystem board_id" on hana and brya

Change-Id: Id69c8e309c0e509a165aa6da2778573ac7de3455
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>