sign_gsc_firmware: check ti50 images for prohibited blobs
We want to add an additional layer of protection against accidental
releasing of prod signed images with dev public keys and hashes for
which private keys are not secret.

The blobs of the keys and hashes to avoid are available in the Ti50
tarball, this patch adds a check and fails the signing process each
time the prohibited blob is found in the Ti50 binary.

BRANCH=none
BUG=b:254059627
TEST=invoked the script to sign Ti50 images built with and without
     'ALLOW_AP_RO_DEV_SIGNING_KEY=1 TI50_DEV=1' defined, Observed
     signer failure when signing the image with either variable
     defined, reporting the presence of the appropriate blob.

(cherry picked from commit 56ca7296a0c35ca8a7b06584d7f714551c918bef)

Change-Id: I8497e749807f862f6d20cf33cad4657008a6372a
Original-Signed-off-by: Vadim Bendebury <vbendeb@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4032539
Original-Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-by: Allen Webb <allenwebb@google.com>
Original-Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-by: Jett Rink <jettrink@chromium.org>
GitOrigin-RevId: 56ca7296a0c35ca8a7b06584d7f714551c918bef