ap_ro_signing: improve key generation, do not sign unless RO_GSCVD present
The path of generating pre-mp AP firmware keys results in an
incomplete AP RO verification keyset, all keys are generated, but the
platform key keyblock is not created. This patch fixes this problem.

Another issue with AP RO verification signing is that
sign_official_build.sh script uses the presence of the AP RO
verification keys as the indicator that AP RO verification signing is
required. But it is possible to have they keys created, but the AP
firmware image still not have the RO_GSCVD section in FMAP.

Using the presence of RO_GSVD section is a more reliable indicator of
the need to sign for AP RO verification. Let's use it and fail the
signer if the section is present, but the AP RO signing keys are not
found in the keys directory.

BRANCH=none
BUG=b:299965578
TEST=generated keys using create_new_keys.sh, attempted to re-sign
     Dedede and Nissa chrome OS images, observed expected behavior (no
     attempt to sigh AP RO on Deded, signed AP RO on Nissa).

     removed the generated arv_root key and tried signing an image
     requiring AP RO verification signing, observed the script
     terminate with error reporting the missing key

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I4ad3272fb62a91154458d3b770b2c91a2beffc5b