ap_ro_signing: do not sign unless RO_GSCVD present
The sign_official_build.sh script uses the presence of the AP RO
verification keys as the indicator that AP RO verification signing is
required. But it is possible to have they keys created, but the AP
firmware image still not have the RO_GSCVD section in FMAP.

Using the presence of RO_GSVD section is a more reliable indicator of
the need to sign for AP RO verification. Let's use it and fail the
signer if the section is present, but the AP RO signing keys are not
found in the keys directory.

BRANCH=none
BUG=b:259965578
TEST=removed the generated arv_root key and tried signing an image
     requiring AP RO verification signing, observed the script
     terminate with error reporting the missing key

(cherry picked from commit 68d9ba190fe8a68f8cdd0fb0e20a38433def1984)

Original-Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I4ad3272fb62a91154458d3b770b2c91a2beffc5b
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4045049
Original-Reviewed-by: Mike Frysinger <vapier@chromium.org>
GitOrigin-RevId: 68d9ba190fe8a68f8cdd0fb0e20a38433def1984
Cr-Build-Id: 8739997580243611025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739997580243611025
Copybot-Job-Name: android-main-fw-nissa-vboot_reference-copybot-downstream