Makefile: Fix missing utilities for 'make all'
The default make target 'all' should depend on either ${UTIL_FILES_SDK}
or ${UTIL_FILES_BOARD}. However, these variables are defined after the
'all' rule, so they actually evaluate to empty strings. This causes the
utility files such as 'crossystem' to be missing. The bug was introduced
by CL:3299836.

Fortunately, these files are also dependencies of the 'install' target.
Therefore this bug doesn't lead to any problem for the vboot_reference
package, because the utilities will be built in src_install().

BUG=none
TEST=make clean && make all -j8 && make install DESTDIR=./install
TEST=make run2tests -j
BRANCH=none

Change-Id: I0abb320e38f17fead373ba44521f3db5346add65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4045570