futility: try ignoring GBB flags when validating GSCVD
GBB flags contents are ignored when AP RO ranges hash is calculated.
The embedded verification will succeed only if the flags are cleared,
but the command line tool should not fail because of nonzero GBB
flags.

This patch adds add additional pass when validating to see if
validation succeeds with GBB flags zeroed.

Also adding a debug printout to allow the user to see ranges covered
by the signature when validating an image and modifying the tests to
accommodate passing when GBB flags are non-zero.

BRANCH=none
BUG=none
TEST=successfully validated AP RO signature with the same image with
     and without cleared gbb flags. When checking the image with
     nonzero flags the 'Ranges digest matches with zeroed GBB flags'
     warning message is printed.

     invoking 'make runtests'  succeeds.

Change-Id: I3e38924f14697a3efd058286f9579d89e5161910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4049934