futility: updater: Only apply the preserve_me quirk for autoupdates.
The preserve_me applies for all non-factory updates for firmware with
the quirk enabled. It was only really intended to apply to firmware
updates during autoupdates, that is --mode=autoupdate. Instead, we
checked for an archive, which is always set, possibly a fallback
directory archive rather than an archive file, resulting in it being
used except for --mode=factory. Switch the condition to TRY_UPDATE_AUTO
instead so only --mode=autoupdate enables the preserve_me quirk.

BUG=b:255447297
TEST=futility update -i /tmp/image.bin doesn't apply the quirk
     futility update -i /tmp/image.bin -m autoupdate applies the quirk
BRANCH=None

Change-Id: I7459f027a918dc70cbde1bfc6f5da2b549bcc513
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4050014