create_new_keys: use MP key for AP RO pubkey block signing
AP RO verification platform keys are signed by the AP RO verification
root key. The GSC which validates the platform key uses hardcoded hash
of the AP RO root key, the same for MP and PreMP images, the same root
key is used for all platforms.

The AP RO verification key is created on the HSM and is stored in
'pem' format, which requires a different vbuitil_key invocation.

With this patch, it is expected that when create_new_keys.sh script is
invoked, the AP RO verification private key directory is accessible to
the signer. The directory is passed through --arv-root-path command
line parameter, by default it is expected to be ./ApRoV1Signing-MP.

BRANCH=none
BUG=b:299965578
TEST=created AP RO root key directory using hsm scripts, then
     successfully invoked create_new_keys.sh

Change-Id: Ic024ccdcdcb751be677934bf559c40b2826c714e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>