firmware/2lib: Introduce API to report previous boot failure
Currently when failures are reported before a slot is selected, vboot
directly requests for recovery. Add a new API to report previous boot
failure before a slot is selected.

BUG=b:242825052
BRANCH=None
TEST=Build Skyrim BIOS image. Run the unit test built for this API. Boot
to OS in Skyrim. Corrupt certain sections in flashmap and report boot
failures and ensured that vboot selected the appropriate FW slot.

Change-Id: I3b1fe8e28fc754919cd4067eeed5029e7dbae7a4
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>