firmware/2lib: Introduce API to report early failure
Currently when failures are reported before a slot is selected, vboot
directly requests for recovery. Add a new API to report early failures
before a slot is selected. This will enable verstage to report failures
observed in components not using verified boot on prior boot cycles.

BUG=b:242825052
BRANCH=None
TEST=Build Skyrim BIOS image. Run the unit test built for this API. Boot
to OS in Skyrim.

Change-Id: I3b1fe8e28fc754919cd4067eeed5029e7dbae7a4
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>