Port sign_uefi.sh to Python
Shell scripts are hard to modify and hard to test, so port sign_uefi.sh
to Python. This is a fairly direct port that attempts to keep all the
behavior the same. In particular, there are no hard errors for missing
EFI/kernel files, or for failing to sign one of those files if it does
exist. It might be good to make the script more strict in the future,
but for now try to match the existing behavior.

Nothing actually calls the new script yet.

Also enable `black_check` in `PRESUBMIT.cfg` to enforce formatting.

BRANCH=none
BUG=b:261631233
TEST=make runtests
TEST=cros lint scripts/image_signing/sign_uefi*.py

(cherry picked from commit 47594a266056d8fc0acc94b3fdf39c261086a0f3)

Change-Id: I4b9b86607cc403779b0504758dd097b0d7237fef
Original-Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4083506
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-by: Mike Frysinger <vapier@chromium.org>
GitOrigin-RevId: 47594a266056d8fc0acc94b3fdf39c261086a0f3