futility: updater: refactor DUT system info retrieval
When developers (or the lab) runs 'futility update' on a Chromebox to
update a remote DUT connected via servo, the updater will incorrectly
recognize the Chromebox as the 'host' = 'system' = 'DUT', selecting
wrong config and setting wrong cookies.

To fix that, we want to isolate and refactor how we identify and access
'host' and 'DUT'. The first step is to rename and move the 'system
property' related functions to 'dut properties' in the `updater_dut.c`.

No functional changes in this patch.
Only renamed functions and moved the implementation to different places.

BUG=b:247428499,b:255617349
TEST=make; run test
BRANCH=None

(cherry picked from commit 9813a9b4470bd1609c853a2381f050faebd4902e)

Change-Id: I5c1f9bb67a14fbcdd80958597290a2789f4c2dac
Original-Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4181581
Original-Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
GitOrigin-RevId: 9813a9b4470bd1609c853a2381f050faebd4902e
Cr-Build-Id: 8739997580243611025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739997580243611025
Copybot-Job-Name: android-main-fw-nissa-vboot_reference-copybot-downstream