futility: updater: refactor DUT system info retrieval
When developers (or the lab) runs 'futility update' on a Chromebox to
update a remote DUT connected via servo, the updater will incorrectly
recognize the Chromebox as the 'host' = 'system' = 'DUT', selecting
wrong config and setting wrong cookies.

To fix that, we want to isolate and refactor how we identify and access
'host' and 'DUT'. The first step is to rename and move the 'system
property' related functions to 'dut properties' in the `updater_dut.c`.

No functional changes in this patch.
Only renamed functions and moved the implementation to different places.

BUG=b:247428499,b:255617349
TEST=make; run test
BRANCH=None

Change-Id: I5c1f9bb67a14fbcdd80958597290a2789f4c2dac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4181581