sign_official_build: Sign reven recovery images like base images
Reven generates recovery images for use with the ChromeOS Recovery
Utility. Since those images are used to install Reven, they should be
signed like base images rather than recovery images (i.e. kern-a/root-a
should be signed with the normal key, not the recovery key).

With our current boot stack this difference doesn't actually matter
since it boots with the kernel image on the ESP, but with crdyboot we'll
boot with the ChromeOS kernel partitions, so they need to be signed
correctly.

BRANCH=none
BUG=b:256621033
TEST=USE=crdyboot build_packages
TEST=build_image
TEST=platform/vboot_reference/scripts/image_signing/sign_official_build.sh
TEST=  recovery build/images/reven/latest/chromiumos_image.bin
TEST=  platform/vboot_reference/tests/devkeys
TEST=  build/images/reven/latest/chromiumos_image.bin.signed
TEST=Boot in a UEFI VM

Change-Id: Ie3f6dd829a8bc15eeffc1d7fe1fc07e14cb30ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4206908