sing_official_build: do not attempt signing AP RO for guybrush
Guybrush firmware archive pinned to ChromeOS firmware releases
starting with 108 includes the uninitialized RO_GSCVD section, which
causes signer failures.

Fixing and qualifying the firmware is going to take a long time, as a
workaround this patch adds a check for the keyset name, and does not
attempt AP RO signing in case the keyset name includes the string
'guybrush', case insensitive.

BRANCH=none
BUG=b:263378945

TEST=downloaded a ChromeOS guybrush recovery image
     hromeOS-recovery-R108-15183.72.0-guybrush.tar.xz, extracted the
     binary and attempted to sign it by running

     ./scripts/image_signing/sign_official_build.sh base \
        <path to>/recovery_image.bin ./test/devkeys /tmp/signed

    observed the expected message on the console:

     "Not looking for RO_GSCVD on guygrush, b/263378945"

   as well as successful firmware singing.

   The same invocation before this patch is applied is failing.

   Also verified that when singing a Nissa recovery image the AP RO
   verification section is signed as expected.

Change-Id: Ifd93e3ebc8fc35e2ba29c7a364767e4414df051d
Signed-off-by: Vadim Bendebury <vbendeb@google.com>