sign_official_build: Support signing KERN-C with different recovery key
Recovery and factory images will now contain a KERN-C, which is
identical to KERN-A except that it may be signed with a different key.
If a new keyblock recovery_kernel.v1.keyblock exists, use that as the
keyblock for KERN-C in the recovery image. Similarly, if
installer_kernel.v1.keyblock exists, use that as the keyblock for KERN-C
in the factory shim. If the new keyblocks don't exist, don't sign
KERN-C. If KERN-C is present in the image, it will remain there signed
with dev keys.

BRANCH=None
BUG=b:266502803
TEST=- Run replace_recovery_key.sh in devkeys directory to get keys for
testing.
- Run sign_official_build.sh on a recovery image with KERN-C present.
- Set recovery_key.vbpubk in GBB - recovery succeeds using KERN-A.
- Set recovery_key.v1.vbpubk in GBB - recovery succeeds using KERN-C.
- Sign a base image and check it boots.
- Sign a factory shim and check it boots (without replacing keys, since
  KERN-C in factory shims is not implemented yet).

Change-Id: I39b209e1efd4669128a12751f1c4ee94bb722d67
Signed-off-by: Reka Norman <rekanorman@google.com>