sign_official_build: Support a second recovery key
Some devices have a second recovery key, which is used to sign:
- a second recovery kernel KERN-C in recovery images
- a second installer kernel KERN-B in factory images

If a device has a second recovery key, use it to sign the second
recovery and installer kernels. Otherwise, don't sign the second
kernels. If they are present, they'll remain in the image signed with
dev keys.

BRANCH=None
BUG=b:266502803
TEST=- Run replace_recovery_key.sh in devkeys directory to get keys for
testing.
- Run sign_official_build.sh on a recovery image with KERN-C present.
- Set recovery_key.vbpubk in GBB - recovery succeeds using KERN-A.
- Set recovery_key.v1.vbpubk in GBB - recovery succeeds using KERN-C.
- Sign a base image and check it boots.
- Sign a factory shim and check it boots (without replacing keys, since
  KERN-C in factory shims is not implemented yet).

Change-Id: I39b209e1efd4669128a12751f1c4ee94bb722d67
Signed-off-by: Reka Norman <rekanorman@google.com>