futility: updater: detect DUT type
Detect if we are going to update a remote DUT (by checking if the flash
parameter has been changed), and ignore all the local system properties
if needed.

This should help 'updating a Chromebook from a Chromebox via servo' to
behave the same as updating from a non-ChromeOS Linux desktop.

BUG=b:247428499,b:255617349
TEST=make; run test
BRANCH=None

Change-Id: I4aa0e98efa21179708d8b593fc619b7f7b65f418
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4181582