futility/cmd_gbb_utility.c: Fix string length off-by-1
Do not copy terminator character to GBB structure to allow 16 byte
HWIDs to be used.

WIP.

Not clear if this is the right fix or if the test should be changed to
only use 15 byte IDs. However, tests/futility/test_gbb_utility.sh has
had a 16 byte ID test since 2022-06-08.

BUG=none
TEST=not yet

Change-Id: If93c92b07d99cae1e9e1befc856da3894009dac9