futility: updater: revise DUT type detection
Some commands like 'cros ap flash' may pass '-p host' to the futility
updater, making the updater consider itself running in the remote DUT
mode.

For other futility commands they work on files by default and do
flashing if '-p' is specified, but that is not the case for the updater,
so we should double check the programmer before deciding the DUT type.

BUG=b:271115449
TEST=FEATURES=test emerge vboot_reference
BRANCH=None

Change-Id: I5be6e4382acffced0ef9cd94f6ede9e9d3fe2efa
Signed-off-by: Hung-Te Lin <hungte@chromium.org>