tpm2_lite: Treat NV_UNINITIALIZED error as BADINDEX for TPM 2.0
This patch ports the equivalent of CB:20299 to the vboot version of the
TPM library. This should fix a problem that was observed where a created
but not written FWMP space gets a device stuck in recovery mode.

BRANCH=none
BUG=b:272310645
TEST=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: If0ef6fdda879b45de0444cdaeb423d01ce109e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4327074
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit 9082d13a642049aae3002850296b5fc53d6b687f)