futility/cmd_create.c: Drop nonsense logic
Here `if (!infile)` branch is always true assuming initialised
memory to be `NULL` and thus implies either branch with,
`if (argc - optind <= 0)` or continue and set the infile
unconditionally `infile = argv[optind++]`.

Therefore, we can conclude that `if (argc > optind)` is always
true which implies `remove_ext` is always false and subsequently
the corresponding logic is spurious.

BUG=none
BRANCH=none
TEST=`cros_run_unit_tests --host --packages vboot_reference`.

Change-Id: I5a11e9e71576297d9bb99ac9cf24aaaa3b030917
Signed-off-by: Edward O'Callaghan <quasisec@google.com>