futility/cmd_flash_util.c: Print fw expected wp sr conf values
AP RO verification in the case of non-factory flows wishes
to determine the WP configuration values needed to calculate
the SR configuration data as defined by the firmware image.

Since we already do this to configure WP correctly on the
DUT using FW derived values parsed from the WP_RO region
within the FMAP we can simply print the offset and length.

BUG=b:281888335
TEST=none

Change-Id: Iafa0cc4193453cc67952aa4ff1f1554ac1861882
Signed-off-by: Edward O'Callaghan <quasisec@google.com>