futility/updater_dut: Fix return value
dut_get_property_string was changed to return an int in CL:4517774, but
on failure it still tries to return NULL. That doesn't fly with some
compilers/flags (e.g. as used in the servod docker container build), so
go for -1 like in dut_set_property_string.

The one user I found for the function compares to 0 for "success", so
returning -1 appropriately reports failure.

BUG=none
TEST=none

Change-Id: I872449a22f3d84fd8bc86125b35779c46dd8c642
Signed-off-by: Patrick Georgi <pgeorgi@google.com>