futility: Follow-up fixes to CL:4548417
This patch syncs futility sign with what was done for vbutil_kernel in
CL:4548417 and makes the --bootloader argument optional. Also make the
--version argument in futility sign optional for kernels which aligns it
better with how it already behaves for firmware, and add some more
checks to reject options that don't apply to repacking existing
preambles. Update tests to reflect the new state.

BUG=none
TEST=make runtests
BRANCH=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ib955e42bcf42327298758aa2891a0703c98bb43b