vbutil_kernel: Drop alignment check for EFI stub
The bootloader size value comes from the setup_sects field in the kernel
params header. The size is 512-aligned, not 4096-aligned, and in some
reven kernel builds the size has been observed to be 0x3800 which is
indeed not 4K aligned and fails this check.

Note that the g_bootloader_size is still rounded up to 4K a few lines
later.

BUG=b:289137837
TEST=cros_workon --host start vboot_reference
TEST=sudo emerge vboot_reference
TEST=USE="-kernel-5_15 kernel-6_1" cros build-packages --board=reven
TEST=build_image
BRANCH=none

Change-Id: If5816033af7e9a8af483fcc9e040cc6a82c8f459
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>