futility: updater: support non-unibuild custom label devices running unibuild
For non-unibuild custom label devices (e.g., hana) being updated to
unibuild, their custom label tag in VPD was still following the old name
(e.g., "customization_id") and we have to fallback to the old VPD
entries for these devices.

One side effect is if there is a unibuild custom label device that
accidentally set "customization_id" and the value is different than the
default fallback (e.g., the model name) then it would fail to update,
but there should be no such devices as far as we know at the moment.

BUG=b:286528262
TEST=make; run unit test
BRANCH=None

Change-Id: Ib3219e2e7526404b290fe4d19c7948267e0b97c9
Signed-off-by: Hung-Te Lin <hungte@chromium.org>