futility: Pull file names into ft_show_bios() subtypes
This patch extends the concept from the previous patch to rename all
`name` arguments to futility show functions to `fname` to prevent
confusion that can lead to bugs. Add a separate `print_name` when
needed. This fixes a bug with metadata hash output when running
`futility show` on the `bios` file type.

BRANCH=none
BUG=none
TEST=make runtests

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I4c7077b0110d3b017242a6c59a4488a0bdef3dd9